requirements.txt
annotated-types==0.5.0
anyio==4.0.0
atproto==0.0.29
backports.pbkdf2==0.1
bases==0.2.1
blis==0.7.8
catalogue==2.0.8
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cohost==0.2.1
colorama==0.4.5
cycler==0.11.0
cymem==2.0.6
dacite==1.8.1
dag-cbor==0.3.2
discord-webhook==0.16.3
exceptiongroup==1.1.3
fastai==2.7.9
fastcore==1.5.17
fastdownload==0.0.7
fastprogress==1.0.3
flickrapi==2.4.0
fonttools==4.35.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.3
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.4
langcodes==3.3.0
libipld==1.0.0
MarkupSafe==2.1.1
matplotlib==3.5.3
multiformats==0.2.1
multiformats-config==0.2.0.post4
murmurhash==1.0.8
numpy==1.23.2
oauthlib==3.2.0
packaging==21.3
pandas==1.4.3
pathy==0.6.2
Pillow==9.2.0
preshed==3.0.7
pydantic==2.3.0
pydantic-core==2.6.3
PyJWT==2.7.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
scikit-learn==1.1.2
scipy==1.9.0
shortuuid==1.0.11
six==1.16.0
smart-open==5.2.1
sniffio==1.3.0
spacy==3.4.1
spacy-legacy==3.0.9
spacy-loggers==1.0.3
srsly==2.4.4
thinc==8.1.0
threadpoolctl==3.1.0
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.0
tweepy==4.14.0
typer==0.4.2
typing-extensions==4.7.1
typing-validation==1.0.0.post2
urllib3==1.26.11
wasabi==0.10.1
websockets==11.0.3
wget==3.2
